Restructure text fields and separate to classes
Instead of passing an attribtue_type into the initialize method make the
various input types inherit from Inputs::Base and have each subclass
define the builder method. This is a cleaner approach, albeit slightly
more-verbose
