+++ lib/govuk_design_system_formbuilder.rb
@@ -4,12 +4,18 @@ require 'govuk_design_system_formbuilder/base'
 
 require 'govuk_design_system_formbuilder/elements/hint'
 require 'govuk_design_system_formbuilder/elements/label'
-require 'govuk_design_system_formbuilder/elements/input'
 require 'govuk_design_system_formbuilder/elements/date'
 require 'govuk_design_system_formbuilder/elements/select'
 require 'govuk_design_system_formbuilder/elements/submit'
 require 'govuk_design_system_formbuilder/elements/text_area'
 require 'govuk_design_system_formbuilder/elements/file'
+
+require 'govuk_design_system_formbuilder/elements/inputs/base'
+require 'govuk_design_system_formbuilder/elements/inputs/email'
+require 'govuk_design_system_formbuilder/elements/inputs/number'
+require 'govuk_design_system_formbuilder/elements/inputs/phone'
+require 'govuk_design_system_formbuilder/elements/inputs/text'
+require 'govuk_design_system_formbuilder/elements/inputs/url'
 
 require 'govuk_design_system_formbuilder/elements/radios/collection'
 require 'govuk_design_system_formbuilder/elements/radios/collection_radio_button'
