+++ lib/govuk_design_system_formbuilder/elements/inputs/email.rb
@@ -1,0 +1,11 @@
+module GOVUKDesignSystemFormBuilder
+  module Elements
+    module Inputs
+      class Email < Base
+        def builder_method
+          :email_field
+        end
+      end
+    end
+  end
+end
