+++ lib/govuk_design_system_formbuilder/elements/inputs/number.rb
@@ -1,0 +1,11 @@
+module GOVUKDesignSystemFormBuilder
+  module Elements
+    module Inputs
+      class Number < Base
+        def builder_method
+          :number_field
+        end
+      end
+    end
+  end
+end
