+++ lib/govuk_design_system_formbuilder/elements/inputs/phone.rb
@@ -1,0 +1,11 @@
+module GOVUKDesignSystemFormBuilder
+  module Elements
+    module Inputs
+      class Phone < Base
+        def builder_method
+          :phone_field
+        end
+      end
+    end
+  end
+end
