+++ lib/govuk_design_system_formbuilder/elements/inputs/text.rb
@@ -1,0 +1,11 @@
+module GOVUKDesignSystemFormBuilder
+  module Elements
+    module Inputs
+      class Text < Base
+        def builder_method
+          :text_field
+        end
+      end
+    end
+  end
+end
