+++ lib/govuk_design_system_formbuilder/elements/inputs/url.rb
@@ -1,0 +1,11 @@
+module GOVUKDesignSystemFormBuilder
+  module Elements
+    module Inputs
+      class URL < Base
+        def builder_method
+          :url_field
+        end
+      end
+    end
+  end
+end
